Move data persistence to IndexedResource
