+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/cloudformation/IndexedResources.kt
@@ -3,26 +3,47 @@
 
 package software.aws.toolkits.jetbrains.services.cloudformation
 
+import java.io.DataInput
+import java.io.DataOutput
+
 /**
  * Immutable data class for indexing [Resource]. Use [from] to create an instance so that it always
  * returns a concrete [IndexedResource] such as [IndexedFunction] if applicable.
  */
-open class IndexedResource internal constructor(val indexedProperties: Map<String, String>) {
+open class IndexedResource protected constructor(val indexedProperties: Map<String, String>) {
 
     protected constructor(resource: Resource, indexProperties: List<String>)
-            : this(indexProperties.map {
-        it to try {
-            resource.getScalarProperty(it)
-        } catch (e: Exception) {
-            ""
+        : this(indexProperties
+            .asSequence()
+            .map { it to try { resource.getScalarProperty(it) } catch (e: Exception) { null } }
+            .mapNotNull { (key, value) -> value?.let { key to it } }
+            .toMap())
+
+    fun save(dataOutput: DataOutput) {
+        dataOutput.writeInt(indexedProperties.size)
+        indexedProperties.forEach { key, value ->
+            dataOutput.writeUTF(key)
+            dataOutput.writeUTF(value)
         }
-    }.toMap())
+    }
 
     override fun equals(other: Any?): Boolean = this === other || (other as? IndexedResource)?.indexedProperties == indexedProperties
 
     override fun hashCode(): Int = indexedProperties.hashCode()
 
     companion object {
+        fun read(dataInput: DataInput): IndexedResource {
+            val mutableMap: MutableMap<String, String> = mutableMapOf()
+
+            val propertySize = dataInput.readInt()
+            repeat(propertySize) {
+                val key = dataInput.readUTF()
+                val value = dataInput.readUTF()
+                mutableMap[key] = value
+            }
+            return IndexedResource(mutableMap)
+        }
+
         fun from(type: String, indexedProperties: Map<String, String>) =
                 INDEXED_RESOURCE_MAPPINGS[type]?.first?.invoke(indexedProperties) ?: IndexedResource(indexedProperties)
 
@@ -37,9 +58,9 @@ class IndexedFunction : IndexedResource {
 
     internal constructor(resource: Resource) : super(resource, listOf("Runtime", "Handler"))
 
-    fun runtime() = indexedProperties["Runtime"]
+    fun runtime(): String? = indexedProperties["Runtime"]
 
-    fun handler() = indexedProperties["Handler"]
+    fun handler(): String? = indexedProperties["Handler"]
 }
 
 internal val INDEXED_RESOURCE_MAPPINGS = mapOf<String, Pair<(Map<String, String>) -> IndexedResource, (Resource) -> IndexedResource>>(
